Rename text_render to background_text_render and simplify
